fix(node10): upgrade chokidar to pass node 10
upath (dependency from chokidar) has failed on many packages because it
was requiring node < 10. Now that upath got updated, chokidar got it too
and we can safely upgrade.

I did not regenerate the package-lock.json file because it was changing
it in a very weird way. I believe it was generated with some nodejs
version I don't have.

I suggest for this to use .nvmrc with a static version so that you can
control the nodejs version used by your contributors and will ease the
process.